Feature/get backend device

* Adding the getBackend, getBackends and getDevice methods

* Installation source directory changed

* Adding the getBackend, getBackends and getDevice to the bindings.

    Changing the aprproximateEntropy binding

* Removing the tsatest.h file

* Renaming test and benchmark header files

* Changing upper to lower case

Approved-by: David Cuesta <david.cuesta@shapelets.io>
Approved-by: Antonio Vilches <antonio.vilches@shapelets.io>

+++ bindings/c/include/tsa_c/features.h
@@ -109,7 +109,7 @@ void cross_covariance(double *xss, long *xss_length, long *xss_number_of_tss, do
  * @param r Filtering level, must be positive.
  * @param result The vectorized approximate entropy for all the input time series in tss.
  */
-void approximate_entropy(double *tss, long *tss_length, long *tss_number_of_tss, int *m, float *r, float *result);
+void approximate_entropy(double *tss, long *tss_length, long *tss_number_of_tss, int *m, double *r, double *result);
 
 #ifdef __cplusplus
 }
